Move vis components into app child directory 'component-library', consolidate dataExporter and jsxExporter into one file.  Jsx is now added directly to componentData, rather than exported when needed for render
