+++ app/component-library/BlackBox.jsx
